design: 달력 아이템 스켈레톤 구현 및 적용
