+++ features/main/components/calendar/atoms/calendar-item.tsx
@@ -1,39 +1,37 @@
 import { useAtomValue } from 'jotai';
 import Link from 'next/link';
-import type { PropsWithChildren } from 'react';
 
 import { calendarDateAtom } from '@/store';
-import { css, cx } from '@/styled-system/css';
-import { flex } from '@/styled-system/patterns';
+import { css } from '@/styled-system/css';
 
 import { Memory } from '../../../types/calendar';
 import { ItemContent } from './calendar-item-content';
+import { CalendarItemSkeleton } from './calendar-item-skeleton';
 
-interface ItemLayoutProps {
+interface CalendarItemProps {
   date: number;
-  isToday: boolean;
-}
-
-interface CalendarItemProps extends ItemLayoutProps {
   memory: Memory | undefined;
   totalDistance?: number;
   isFuture: boolean;
   isDisableRecord: boolean;
+  isFetching: boolean;
 }
 
 export const CalendarItem = ({
   date,
-  isToday,
   isFuture,
   memory,
   isDisableRecord,
+  isFetching,
 }: CalendarItemProps) => {
   const { year, month } = useAtomValue(calendarDateAtom);
   const targetDate = `${year}-${month < 10 ? '0' : ''}${month}-${date < 10 ? '0' : ''}${date}`;
 
+  if (isFetching) return <CalendarItemSkeleton />;
+
   if (!memory)
     return (
-      <ItemLayout date={date} isToday={isToday}>
+      <>
         {isFuture || isDisableRecord ? (
           <div className={linkContainerStyles} />
         ) : (
@@ -42,14 +40,15 @@ export const CalendarItem = ({
             className={linkContainerStyles}
           />
         )}
-      </ItemLayout>
+        <div className={totalDistanceAreaStyles} />
+      </>
     );
 
   const { memoryId, type, totalDistance, strokes, isAchieved, imageUrl } =
     memory;
 
   return (
-    <ItemLayout date={date} isToday={isToday}>
+    <>
       <Link href={`/record-detail/${memoryId}`} className={linkContainerStyles}>
         {
           <ItemContent
@@ -61,60 +60,16 @@ export const CalendarItem = ({
           />
         }
       </Link>
-      {totalDistance && <p>{totalDistance}</p>}{' '}
-    </ItemLayout>
+      <div className={totalDistanceAreaStyles}>
+        {totalDistance && <p>{totalDistance}</p>}
+      </div>
+    </>
   );
 };
 
-const ItemLayout = ({
-  date,
-  isToday,
-  children,
-}: PropsWithChildren<ItemLayoutProps>) => {
-  return (
-    <li className={itemContainerStyles}>
-      <p
-        className={cx(
-          DateStyles,
-          isToday === true ? TodayDateStyles : EmptyStyles,
-        )}
-      >
-        {date}
-      </p>
-      {children}
-    </li>
-  );
-};
+export const totalDistanceAreaStyles = css({ height: '18px' });
 
-const itemContainerStyles = flex({
-  width: 'full',
-  height: 'fit-content',
-  flexDir: 'column',
-  alignItems: 'center',
-
-  '& > p': {
-    height: '18px',
-    textStyle: 'caption2',
-    fontWeight: 'medium',
-    textAlign: 'center',
-  },
-});
-
-const DateStyles = css({
-  padding: '2px',
-  width: '18px',
-  marginBottom: '5px',
-  borderRadius: 'full',
-});
-
-const EmptyStyles = '';
-
-const TodayDateStyles = css({
-  backgroundColor: 'blue.60',
-  color: 'white',
-});
-
-const linkContainerStyles = css({
+export const linkContainerStyles = css({
   width: 'full',
   aspectRatio: 'auto 3 / 4',
   maxHeight: '120px',
