+++ features/main/components/calendar/atoms/index.ts
@@ -1,5 +1,6 @@
 export * from './calendar-item';
 export * from './calendar-item-content';
+export * from './calendar-item-skeleton';
 export * from './day-labels';
 export * from './move-month-button';
 export * from './view-image-button';
